Refactor session management: implement session creation, saving, and logout functionality; save session file to player file

+++ app.py
@@ -7,9 +7,21 @@ def is_authenticated():
 
 def create_session(player_data):
 
-    with open("session.json", "w") as session_file:
-        json.dump(player_data, session_file, indent=1 )
-        print("Creating session...")
+    # Create JSON file if it doesn't exist
+    if not os.path.exists("session.json"):
+        with open("session.json", "w") as session_file:
+            json.dump(player_data, session_file, indent=1 )
+            print("Creating session...")
+
+def save_session(session_data, player_data):
+
+    players = player_data["player"]
+
+    for i,player in enumerate(players):
+        if player["id"] == int(session_data["id"]):
+            player["coin"] = int(session_data["coin"])
+            save_coin(player_data)
+            print("session saved!")
 
 def clear_session():
     if os.path.exists("session.json"):
@@ -33,7 +45,9 @@ def login(username, password, player_data):
     else:
         print("Your username or password are wrong")
 
-def logout():
+def logout(session_data,player_data):
+
+    save_session(session_data, player_data)
     clear_session()
     print("Logged out successfully")
 
@@ -42,10 +56,14 @@ def save_data(data):
     with open("data.json", "w") as file :
         json.dump(data, file, indent=4)
 
-#function to save player.json
 def save_coin(player_data):
     with open("player.json", "w") as file :
         json.dump(player_data, file, indent=4)
+
+def get_player_id (session_data):
+
+    player_id = session_data["id"]
+    return player_id
 
 def add_coin(player_data, player_id):
 
@@ -105,8 +123,6 @@ def display_coin(player_data, player_id):
 
 def sedekah(player_data, player_id, receiver_id):
 
-
-
     players = player_data["player"]
     player_found = False
     receiver_found = True
@@ -134,7 +150,6 @@ def sedekah(player_data, player_id, receiver_id):
     if receiver_found == False:
         print(f"Receiver Player [ID:{receiver_id}] not found")        
 
-
 def main():
 
     # Create JSON file if it doesn't exist
@@ -154,6 +169,12 @@ def main():
     # Open the file
     with open("player.json", "r") as file :
         player_data = json.load(file)
+
+    # Open the file
+    if os.path.exists("session.json"):
+        with open("session.json", "r") as session_file :
+            session_data = json.load(session_file)
+
 
 
     parser = argparse.ArgumentParser()
@@ -176,38 +197,34 @@ def main():
 
     # Display coin parser
     display_coin_parser = subparser.add_parser("coin",help="display your current coin")
-    display_coin_parser.add_argument("player_id", help="your player id", type=int)
 
     # Sedekah parser
     sedekah_parser = subparser.add_parser("sedekah", help="sedekah your coin!")
-    sedekah_parser.add_argument("player_id", help="your player id", type=int)
     sedekah_parser.add_argument("-to", "--to_receiver", dest="receiver_id", help=" the id of receiver", type=int, required=True)
 
     args = parser.parse_args()
 
-   
-
     if args.command == "login":
-        player_id = login(args.username, args.password, player_data)
-        print(player_id)
+        login(args.username, args.password, player_data)
     elif args.command == "logout":
-        logout()
+        logout(session_data, player_data)
     else: 
         if not is_authenticated():
             print("You must login first!")
-            return
-
+            return      
+        
         if args.command == "mark-done":
             deed_id = args.deed_id 
+            player_id = get_player_id(session_data)
             mark_done(data,deed_id,player_data,player_id)
             print("Congrats! Your marking is done")
         if args.command == "list":
             list_deeds(data)
         if args.command == "coin":
-            player_id = args.player_id
+            player_id = get_player_id(session_data)
             display_coin(player_data, player_id)
         if args.command == "sedekah":
-            player_id = args.player_id
+            player_id = get_player_id(session_data)
             receiver_id = args.receiver_id
             sedekah(player_data, player_id, receiver_id)
